Unify the constructors for the backend and frontend spoonform
